Import show-merged-branches script
This script is designed to work around the fact that in GitLab and
GitHub there is an option to "Squash and Merge" a Pull/Merge Request.
When done, the standard git tooling for checking if a branch has been
merged fails to work correctly.  This script can correctly identify
those branches, and also clean up any that are left around if you want
it to.

Signed-off-by: Dan Kalowsky <dkalowsky@amperecomputing.com>